requirements.txt
anndata==0.11.3
conch==8.1.0
einops==0.8.0
einops_exts==0.0.4
environs==14.1.0
h5py==3.11.0
huggingface_hub==0.22.2
loguru==0.7.3
numpy==1.24.4
pandas==2.0.3
peft==0.10.0
Pillow==11.1.0
pytorch_lightning==2.5.0.post0
PyYAML==6.0.1
PyYAML==6.0.2
safetensors==0.5.2
scanpy==1.10.4
scikit_learn==1.3.2
scipy==1.15.1
scprep==1.2.3
timm==0.5.4
torch==2.2.2
torch_geometric==2.6.1
torchmetrics==1.6.1
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
xformers==0.0.29.post1
